refactor: migrate ldap to rest-api-client
