+++ packages/frontend/@n8n/rest-api-client/src/api/index.ts
@@ -3,6 +3,7 @@ export * from './communityNodes';
 export * from './ctas';
 export * from './eventbus.ee';
 export * from './events';
+export * from './ldap';
 export * from './mfa';
 export * from './nodeTypes';
 export * from './npsSurvey';
